Added additional struct for response. Cleaned up Model Objects.
