+++ weatherApp/weatherApp/ViewController.swift
@@ -13,19 +13,8 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view, typically from a nib.
-        let myData = JSON()
-        
-        let decoder = JSONDecoder()
-        do {
-            
-            let model = try decoder.decode(WeatherResponse2.self, from: myData.data)
-            print(model.list.first ?? "nothing to print")
-            
-        } catch {
-            print(error)
-        }
+       
     }
-
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
