+++ weatherApp/weatherApp/WeatherResponse.swift
@@ -2,87 +2,130 @@
 //  WeatherResponse.swift
 //  weatherApp
 //
-//  Created by Lewis McGrath on 13/11/2017.
+//  Created by Lewis McGrath on 15/11/2017.
 //  Copyright © 2017 Lewis McGrath. All rights reserved.
 //
 
 import Foundation
 
-//Build out the model objects based on expected JSON
-
-//Think this might be wrong?
-struct BaseResponse {
-    let weatherResponse : [WeatherMainInfo]
-    let dataText : String
+struct WeatherResponse: Codable {
     
+    private enum CodingKeys: String, CodingKey {
+        case count = "cnt"
+        case list = "list"
+        case city = "city"
+    }
     
-//"List" marker (array) containts these JSON Objects
-    struct WeatherMainInfo {
-        let temp : Double
-        let tempMin : Double
-        let tempMax : Double
-        let pressure : Double
-        let seaLevel : Double
-        let grndLevel : Double
-        let humidity : Double
-        let tempKf : Double
+    let count: Float
+    let list: [WeatherList]
+    let city: City
+}
+
+struct City: Codable {
+    private enum CodingKeys: String, CodingKey{
+        case id = "id"
+        case name = "name"
+        case coordinates = "coord"
+        case country = "country"
     }
+    let id : Int
+    let name : String
+    let coordinates : Coord
+    let country : String
+}
 
-//TODO: Fix. Object with an inner array object
-    struct Weather {
-        let id : Int
-        let main : String
-        let description : String
-        let icon : String
-    }
-
-    struct Clouds {
-        let all : Int
-    }
-
-    struct Wind {
-        let speed : Double
-        let deg : Double
-    }
-
-    struct sys {
-        let pod : String
-        }
-    }
-
-// Set url that will be used for API call
-// Set up request constant that makes a URL Request with the url constant
-// Set up the URLsession with a data task and a completion handler
-// Set up the task
-
-let url = "http://api.openweathermap.org/data/2.5/forecast?q=London&APPID=296df7e264a3493489577e11f60b9c6b"
-
-let request = URLRequest(url: URL(string: url)!)
-
-let task = URLSession.shared.dataTask(with: request) { (data, response, error) in
+struct Coord: Codable {
     
-    guard let data = data else { return }
+    private enum CodingKeys: String, CodingKey{
+        case latitude = "lat"
+        case longitude = "lon"
+    }
+    
+    let latitude: Float
+    let longitude: Float
+}
 
-   let jsonDecoder = JSONDecoder()
+struct WeatherList: Codable {
+    
+    private enum CodingKeys: String, CodingKey {
+        case main = "main"
+        case weather = "weather"
+        case clouds = "clouds"
+        case wind = "wind"
+        case sys = "sys"
+        case dateText = "dt_txt"
+    }
+    
+    let main: MainWeather
+    let weather: [Weather]
+    let clouds: Clouds
+    let wind: Wind
+    let sys: Sys
+    let dateText : String
+}
 
+struct MainWeather: Codable {
+    
+    private enum CodingKeys: String, CodingKey {
+        case temp = "temp"
+        case tempMin = "temp_min"
+        case tempMax = "temp_max"
+        case pressure = "pressure"
+        case seaLevel = "sea_level"
+        case groundLevel = "grnd_level"
+        case humidity = "humidity"
+        case tempKf = "temp_kf"
+    }
+    
+    let temp: Float
+    let tempMin: Float
+    let tempMax: Float
+    let pressure: Float
+    let seaLevel: Float
+    let groundLevel : Float
+    let humidity: Float
+    let tempKf : Float
+}
+
+struct Weather: Codable {
+    let id: Float
+    let main: String
+    let description: String
+    let icon: String
+}
+
+struct Clouds: Codable {
+    let all: Float
+}
+
+struct Wind: Codable {
+    let speed: Float
+    let deg : Float
+}
+
+struct Sys: Codable{
+    let pod : String
 }
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+    let url = "http://api.openweathermap.org/data/2.5/forecast?q=London&APPID=296df7e264a3493489577e11f60b9c6b"
+    
+    let request = URLRequest(url: URL(string: url)!)
+    
+    let task = URLSession.shared.dataTask(with: request) { (data, response, error) in
+        
+        guard let myData = data else { return }
+        
+        let jsonDecoder = JSONDecoder()
+        do {
+            
+            let model = try jsonDecoder.decode(WeatherResponse.self, from: myData)
+            print(model.list.first ?? "There is nothing to print!")
+            
+        } catch {
+            print(error)
+        }
+}
